requirements.txt
altair==5.3.0
attrs==23.2.0
azure-ai-textanalytics==5.3.0
azure-common==1.1.28
azure-core==1.30.2
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
inexactsearch==1.0.2
isodate==0.6.1
jellyfish==1.0.4
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
networkx==3.3
nltk==3.8.1
numpy==1.26.4
packaging==24.1
pandas==2.2.2
pillow==10.3.0
protobuf==4.25.3
pyarrow==16.1.0
pydeck==0.9.1
Pygments==2.18.0
pyspellchecker==0.8.1
python-dateutil==2.9.0.post0
pytz==2024.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
scikit-learn==1.5.0
scipy==1.13.1
segtok==1.5.11
setuptools==70.1.0
silpa_common==0.3
six==1.16.0
smmap==5.0.1
soundex==1.1.3
spellchecker==0.4
streamlit==1.35.0
tabulate==0.9.0
tenacity==8.4.1
threadpoolctl==3.5.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.1
yake==0.4.8
streamlit
nltk
pandas
sqlite3
streamlit
nltk
pandas
yake

